// client/src/open_api/model/profile.ts
/**
 * API V1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProfileAddress } from './profileAddress';


export interface Profile { 
    first_name: string;
    last_name: string;
    address: ProfileAddress;
}